+++ src/oxl_ansible_webui/aw/execute/queue.py
@@ -1,5 +1,6 @@
 from aw.model.job import JobExecution, JobQueue
-from aw.utils.debug import log
+from aw.utils.debug import log, log_security
+from aw.utils.crypto import decrypt, encrypt
 
 
 def queue_get() -> (JobExecution, None):
@@ -8,10 +9,16 @@ def queue_get() -> (JobExecution, None):
         return None
 
     execution = next_queue_item.execution
+    v = next_queue_item.v
     next_queue_item.delete()
+
+    if decrypt(v) != str(execution):
+        log_security(f"Job '{execution.job.name} {execution.id}' failed execution queue validation - possible attack")
+        return None
+
     return execution
 
 
 def queue_add(execution: JobExecution):
     log(msg=f"Job '{execution.job.name} {execution.id}' added to execution queue", level=4)
-    JobQueue(execution=execution).save()
+    JobQueue(execution=execution, v=encrypt(str(execution))).save()